Move keying_material_len to generic srtp
