+++ src/crypto/srtp.rs
@@ -11,6 +11,26 @@ pub enum SrtpProfile {
     AeadAes128Gcm,
 }
 
+impl SrtpProfile {
+    // All the profiles we support, ordered from most preferred to least.
+    pub(crate) const ALL: &'static [SrtpProfile] =
+        &[SrtpProfile::AeadAes128Gcm, SrtpProfile::Aes128CmSha1_80];
+
+    /// The length of keying material to extract from the DTLS session in bytes.
+    #[rustfmt::skip]
+    pub(crate) fn keying_material_len(&self) -> usize {
+        match self {
+            #[cfg(feature = "_internal_test_exports")]
+            SrtpProfile::PassThrough => 0,
+             // MASTER_KEY_LEN * 2 + MASTER_SALT * 2
+             // TODO: This is a duplication of info that is held in srtp.rs, because we
+             // don't want a dependency in that direction.
+            SrtpProfile::Aes128CmSha1_80 => 16 * 2 + 14 * 2,
+            SrtpProfile::AeadAes128Gcm   => 16 * 2 + 12 * 2,
+        }
+    }
+}
+
 #[allow(unused)]
 
 pub fn new_aes_128_cm_sha1_80(
